wip, refactoring a bit, using chi web server router, thinking about how to organize better
